Move `TemplateString` to a separate file
