+++ java-runtime/src/main/proto/spine/validate/validation_error.proto
@@ -36,6 +36,7 @@ option java_package = "io.spine.validate";
 
 import "google/protobuf/any.proto";
 
+import "spine/validate/error_message.proto";
 import "spine/base/field_path.proto";
 
 // An error indicating that a message did not pass validation.
@@ -59,7 +60,7 @@ message ConstraintViolation {
     repeated string param = 2 [deprecated = true];
 
     // An error message for this violation.
-    TemplateString message = 8;
+    spine.validate.TemplateString message = 8;
 
     // The name of the validated message type.
     string type_name = 7;
@@ -81,36 +82,4 @@ message ConstraintViolation {
     // validation constraints, and simply non-default value was required in the parent message.
     //
     repeated ConstraintViolation violation = 5;
-}
-
-// Represents a template string with placeholders and a map for further substituting
-// those placeholders with the actual values.
-//
-// Placeholders are specified using the format `${key}`, where `key` is the identifier
-// for the value to be substituted from the `placeholder_value` map. There are no
-// expectations about the identifier format. It can be `myKey`, `my.key`, `my_key`, etc.
-//
-// Example usage:
-//     template_string = "My dog's name is ${dog.name}."
-//     placeholder_values = { "dog.name": "Fido" }
-//
-// After substitution, the final output would be:
-//     "My dog's name is Fido."
-//
-// Each placeholder `key` should have a corresponding entry in `placeholder_value`.
-//
-message TemplateString {
-
-    // The template string that may contain one or more placeholders.
-    string with_placeholders = 1;
-
-    // A map that provides values for each placeholder referenced in `with_placeholders`.
-    //
-    // The keys in this map should match the placeholder keys inside `with_placeholders` without
-    // the `${}` prefix.
-    //
-    // All placeholders present in `with_placeholders` must have corresponding entries in this map,
-    // otherwise the template is considered invalid.
-    //
-    map<string, string> placeholder_value = 2;
 }
