+++ java-tests/runtime/src/test/kotlin/io/spine/validate/EnclosedMessageValidationSpec.kt
@@ -40,7 +40,6 @@ import io.spine.test.validate.notValidateEnclosed
 import io.spine.validate.ValidationOfConstraintTest.Companion.VALIDATION_SHOULD
 import io.spine.validate.given.MessageValidatorTestEnv.EMAIL
 import io.spine.validate.given.MessageValidatorTestEnv.ENCLOSED_FIELD_NAME
-import io.spine.validate.text.format
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
