+++ java-tests/validating/src/test/kotlin/io/spine/test/options/NumberConstraintsITest.kt
@@ -33,7 +33,7 @@ import io.kotest.matchers.string.shouldContain
 import io.spine.test.tools.validate.InterestRate
 import io.spine.test.tools.validate.Probability
 import io.spine.test.tools.validate.Year
-import io.spine.validate.text.format
+import io.spine.validate.format
 import io.spine.validation.assertions.assertInvalid
 import io.spine.validation.assertions.assertValid
 import org.junit.jupiter.api.DisplayName
