+++ java-tests/vanilla/src/test/kotlin/io/spine/validation/java/IsRequiredSpec.kt
@@ -31,7 +31,7 @@ import com.google.protobuf.Message
 import io.spine.testing.TestValues.randomString
 import io.spine.validate.NonValidated
 import io.spine.validate.Validate.violationsOf
-import io.spine.validate.text.format
+import io.spine.validate.format
 import io.spine.validation.java.given.Fish
 import io.spine.validation.java.given.Meal
 import io.spine.validation.java.given.Sauce
